fix: fixed a great number of compiler warnings :)

+++ trunk/datinfodbmgr.cpp
@@ -383,7 +383,7 @@ void DatInfoDatabaseManager::setCacheSize(quint64 kiloBytes)
 void DatInfoDatabaseManager::setSyncMode(uint syncMode)
 {
 	static QStringList dbSyncModes = QStringList() << "OFF" << "NORMAL" << "FULL";
-	if ( syncMode > dbSyncModes.count() - 1 )
+	if ( (int)syncMode > dbSyncModes.count() - 1 )
 		return;
 	QSqlQuery query(m_db);
 	if ( !query.exec(QString("PRAGMA synchronous = %1").arg(dbSyncModes[syncMode])) )
@@ -393,7 +393,7 @@ void DatInfoDatabaseManager::setSyncMode(uint syncMode)
 void DatInfoDatabaseManager::setJournalMode(uint journalMode)
 {
 	static QStringList dbJournalModes = QStringList() << "DELETE" << "TRUNCATE" << "PERSIST" << "MEMORY" << "WAL" << "OFF";
-	if ( journalMode > dbJournalModes.count() - 1 )
+	if ( (int)journalMode > dbJournalModes.count() - 1 )
 		return;
 	QSqlQuery query(m_db);
 	if ( !query.exec(QString("PRAGMA journal_mode = %1").arg(dbJournalModes[journalMode])) )
